[python] Move download utilities to a module
